ci: add checkout action
The default behavior of actions/checkout on PRs is performing a merge
commit against the base branch (i.e. master) and checking out the
resulting commit. This is incredibly problematic because you don't
normally do this locally, usually resulting in abnormal lint/test
failures in CI, but not on your local.

Since this behavior of actions/checkout is undesirable, the new
ory/ci/checkout action wraps around actions/checkout, but instead
checks out the HEAD SHA if it's run in a PR. Default actions/checkout
behavior is followed otherwise.